requirements.txt
# MTG Deck Constructor App - Dependencias

# Graphical interface (included in standard Python)
# tkinter - incluido en Python

# Manejo de solicitudes HTTP
requests>=2.28.0

# Image handling
Pillow>=9.0.0

# Data analysis (optional, for advanced statistics)
pandas>=1.5.0

# Data validation
pydantic>=1.10.0

# Logging avanzado (opcional)
loguru>=0.6.0

# Testing (desarrollo)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
coverage>=7.0.0
factory-boy>=3.2.0

# Code formatting (development)
black>=22.0.0
flake8>=5.0.0

# Documentation (development)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0